Rename to `ModelEditor`
